+++ src/telnet.ts
@@ -61,7 +61,7 @@ dns.lookup(host, (err, addr, family) => {
                                 console.error(err.statusCode, err.statusMessage)
                             else
                                 console.error(err.name, err.code)
-                            console.log(`Perhaps DDnet is not running, try: npm run serve &`)
+                            console.log(`Perhaps DDnet is not running, try: npm run net &`)
                             console.log(` or run DDplay as standalone, try: npm run play`)
                             process.exit()
                         })